grpc_greeter_server/requirements.txt
grpcio
protobuf
prometheus_client
cassandra-driver